backend_with_docker/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-dotenv==1.0.0
openai>=1.6.1
starlette==0.27.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
langchain>=0.0.350
langchain-openai==0.0.2
faiss-cpu==1.7.4
pydantic==2.5.2
itsdangerous==2.1.2
bcrypt==4.0.1